refactor(app): UpdateApp and Alerts
