+++ app/src/App/__tests__/DesktopApp.test.tsx
@@ -13,7 +13,7 @@ import { ProtocolsLanding } from '/app/pages/Desktop/Protocols/ProtocolsLanding'
 import { ProtocolRunDetails } from '/app/pages/Desktop/Devices/ProtocolRunDetails'
 import { RobotSettings } from '/app/pages/Desktop/Devices/RobotSettings'
 import { GeneralSettings } from '/app/pages/Desktop/AppSettings/GeneralSettings'
-import { AlertsModal } from '/app/organisms/Alerts/AlertsModal'
+import { AlertsModal } from '/app/organisms/Desktop/Alerts/AlertsModal'
 import { useFeatureFlag } from '/app/redux/config'
 import { useIsFlex } from '/app/redux-resources/robots'
 import { ProtocolTimeline } from '/app/pages/Desktop/Protocols/ProtocolDetails/ProtocolTimeline'
