+++ app/src/organisms/Desktop/Devices/RobotSettings/UpdateBuildroot/ViewUpdateModal.tsx
@@ -11,7 +11,7 @@ import {
 } from '/app/redux/robot-update'
 import { getAvailableShellUpdate } from '/app/redux/shell'
 import { getTopPortalEl } from '/app/App/portal'
-import { UpdateAppModal } from '/app/organisms/UpdateAppModal'
+import { UpdateAppModal } from '/app/organisms/Desktop/UpdateAppModal'
 import { MigrationWarningModal } from './MigrationWarningModal'
 import { UpdateRobotModal } from './UpdateRobotModal'
